sctp: fix fetching port number before binding
SCTP_STREAM test was broken after bcd199bddc18 ("Bind explicitly only if
we need to.") because it was then calling getsockname() before any
actual bind call, leading it to return 0 for port number, which was sent
to the client.

This fixes it by moving the getsockname() call to after the listen(), as
it's already done for other test cases here.