configure.ac: use AC_USE_SYSTEM_EXTENSIONS
functions like splice(2) and sched_setaffinity(2), used
in netperf, require _GNU_SOURCE to be declared. without
it, we see warnings about missing definitions. once
upon a time, AC_GNU_SOURCE would have pulled this in
where appropriate, but that's just an alias these days
for AC_USE_SYSTEM_EXTENSIONS. with this added, we no
longer see the warnings on linux.