netserver: fix permissions change on /dev/null
The netserver program, when run as root, will change the permissions on
/dev/null unconditionally. This is undesireable at best, and horrible at
worst because so much on a system depends on /dev/null to work correctly.

The fix for /dev/null permissions looks like:
sudo chmod a+w /dev/null

Fix the netserver issue by checking to see if the file being used for debug
is /dev/null or an actual file, and only change the permissions if the file
is an actual file.

Signed-off-by: Jesse Brandeburg <jesse.brandeburg@intel.com>